Fix the type used to hold nanoTime, was long should be long long.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@1302421 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/util/concurrent/locks/AbstractQueuedSynchronizer.cpp
@@ -749,7 +749,7 @@ namespace locks {
                         LockSupport::parkNanos(nanosTimeout);
                     }
 
-                    long now = System::nanoTime();
+                    long long now = System::nanoTime();
                     nanosTimeout -= now - lastTime;
                     lastTime = now;
 
@@ -1053,7 +1053,7 @@ namespace locks {
                     break;
                 }
 
-                long now = System::nanoTime();
+                long long now = System::nanoTime();
                 nanosTimeout -= now - lastTime;
                 lastTime = now;
             }
